Refactor mock-async tests to support sync as well
Move the MockCluster code to its own module and move
tests to associated `test_cluster` and `test_async_cluster`
modules.

Added sync retry test, which uncovered issue with sync retries
retrying one less times than the number of retries specified,
so fixed that.
